+++ Week3-Projects/VisitorManagementApplication/VisitorService/app/models/db/EmployeeTable.scala
@@ -4,7 +4,7 @@ import models.request.Employee
 import slick.jdbc.MySQLProfile.api._
 
 class EmployeeTable(tag: Tag) extends Table[Employee](tag, "employees") {
-  def employeeId = column[Option[Long]]("employee_id", O.PrimaryKey, O.AutoInc)
+  def employeeId = column[Option[Int]]("employee_id", O.PrimaryKey, O.AutoInc)
   def employeeName = column[String]("employee_name")
   def organisation = column[String]("organisation")
   def building = column[String]("building")
